# public/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/v1/crewMembers": {
            "get": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Get a list of crew members",
                "operationId": "2cae5a7f13689922c000a3fc7c990c9b",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "The role of the crew member",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/CrewMemberRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CrewMember"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Out of range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Out of range"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Add a new crew member",
                "operationId": "4966672b2daafa0f0596d6c275cb3cf1",
                "requestBody": {
                    "description": "Crew member data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrewMember"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Crew member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created Crew Member"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can add crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can add crew members"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "208": {
                        "description": "Crew Member already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew Member already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/crewMembers/{id}": {
            "get": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Get a specific crew member by ID",
                "operationId": "628aecd4d12f2d12d9a618f516bb29a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the crew member",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crew member details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrewMember"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Crew member not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew member not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Update a specific crew member by ID",
                "operationId": "cefbcb52866c9ed5106a863c77258e5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the crew member",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Crew member data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrewMember"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crew member updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrewMember"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Crew member not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew member not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Crew Members"
                ],
                "summary": "Delete a specific crew member",
                "operationId": "9b0a7c48c165f8e072b8ced8c264d4e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the crew member",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Crew member deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted Crew Member"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Crew member not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crew member not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Only admins can delete crew members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can delete crew members"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CrewMember": {
                "title": "CrewMember model",
                "description": "CrewMember model",
                "properties": {
                    "id": {
                        "description": "ID of the crew member",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "description": "Name of the crew member",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "photo": {
                        "description": "Photo of the crew member",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "role": {
                        "description": "Role of the crew member",
                        "type": "string",
                        "example": "Director"
                    }
                },
                "type": "object"
            },
...
            "CrewMemberRole": {
                "title": "CrewMemberRole",
                "description": "CrewMemberRole enum",
                "properties": {
                    "Actor": {
                        "description": "Represents an actor",
                        "type": "object",
                        "enum": [
                            "actor"
                        ]
                    },
                    "Director": {
                        "description": "Represents a director",
                        "type": "object",
                        "enum": [
                            "director"
                        ]
                    }
                },
                "type": "string",
                "enum": [
                    "actor",
                    "director"
                ],
                "example": "actor"
            }
        }
    }
}